refactor: mv error types
