+++ packages/encryption/src/encryption/EncryptionTypes.ts
@@ -86,17 +86,3 @@ export interface EncryptionResultE2E {
     /** This feature is supported since v37. */
     ephemeralPublicKey?: EC_Public_CryptoKey
 }
-export enum EncryptErrorReasons {
-    ComplexTypedMessageNotSupportedInPayload38 = '[@masknet/encryption] Complex TypedMessage is not supported in payload v38.',
-    PublicKeyNotFound = '[@masknet/encryption] Target public key not found.',
-    AESKeyUsageError = "[@masknet/encryption] AES key generated by IO doesn't have the correct usages or extractable property.",
-}
-export class EncryptError extends Error {
-    static Reasons = EncryptErrorReasons
-    constructor(
-        public override message: EncryptErrorReasons,
-        cause?: any,
-    ) {
-        super(message, { cause })
-    }
-}
