+++ packages/mask/content-script/components/InjectedComponents/DecryptedPost/DecryptedPost.tsx
@@ -1,6 +1,6 @@
 import { Fragment, useContext, useEffect, useReducer } from 'react'
 import { extractTextFromTypedMessage, isTypedMessageEqual, type TypedMessage } from '@masknet/typed-message'
-import type { ProfileIdentifier } from '@masknet/shared-base'
+import { DecryptProgressKind, type ProfileIdentifier } from '@masknet/shared-base'
 
 import Services, { GeneratorServices } from '#services'
 import type { DecryptionProgress, FailureDecryption, SuccessDecryption } from './types.js'
@@ -10,7 +10,7 @@ import { DecryptPostFailed } from './DecryptPostFailed.js'
 import { encodeArrayBuffer, safeUnreachable } from '@masknet/kit'
 import { activatedSiteAdaptorUI } from '../../../site-adaptor-infra/index.js'
 import type { DecryptionContext, EncodedPayload } from '../../../../background/services/crypto/decryption.js'
-import { DecryptIntermediateProgressKind, DecryptProgressKind } from '@masknet/encryption'
+import { DecryptIntermediateProgressKind } from '@masknet/encryption'
 import { type PostContext, usePostInfoDetails, PostInfoContext } from '@masknet/plugin-infra/content-script'
 import { Some } from 'ts-results-es'
 import { uniqWith } from 'lodash-es'
