+++ packages/shared-base/src/serializer/index.ts
@@ -2,16 +2,16 @@ import { Typeson, TypesonPromise } from 'typeson'
 import type { Serialization } from 'async-call-rpc'
 import { Err, None, Ok, Some } from 'ts-results-es'
 import * as BN from 'bignumber.js'
-import { EncryptError, DecryptError } from '@masknet/encryption'
+import { EncryptError, DecryptError, Identifier } from '@masknet/base'
 import { MaskEthereumProviderRpcError } from '@masknet/sdk'
 
 import { blob, builtin, file, filelist, imagebitmap, specialNumbers } from 'typeson-registry'
-import { Identifier } from '@masknet/base'
 import { responseRegedit } from './response.js'
 import { requestRegedit } from './request.js'
 
 const pendingRegister = new Set<() => void>()
 let typeson: Typeson | undefined
+
 function setup() {
     const { default: BigNumber } = BN
     // https://github.com/dfahlander/typeson-registry/issues/27
